Split connection msgs
