+++ modules/src/ics03_connection/handler/conn_open_ack.rs
@@ -5,7 +5,7 @@ use crate::ics03_connection::error::{Error, Kind};
 use crate::ics03_connection::handler::verify::{check_client_consensus_height, verify_proofs};
 use crate::ics03_connection::handler::ConnectionEvent::ConnOpenAck;
 use crate::ics03_connection::handler::ConnectionResult;
-use crate::ics03_connection::msgs::MsgConnectionOpenAck;
+use crate::ics03_connection::msgs::conn_open_ack::MsgConnectionOpenAck;
 
 /// Protocol logic specific to processing ICS3 messages of type `MsgConnectionOpenAck`.
 pub(crate) fn process(
@@ -86,8 +86,9 @@ mod tests {
     use crate::ics03_connection::context::ConnectionReader;
     use crate::ics03_connection::context_mock::MockConnectionContext;
     use crate::ics03_connection::handler::{dispatch, ConnectionResult};
-    use crate::ics03_connection::msgs::test_util::get_dummy_msg_conn_open_ack;
-    use crate::ics03_connection::msgs::{ConnectionMsg, MsgConnectionOpenAck};
+    use crate::ics03_connection::msgs::conn_open_ack::test_util::get_dummy_msg_conn_open_ack;
+    use crate::ics03_connection::msgs::conn_open_ack::MsgConnectionOpenAck;
+    use crate::ics03_connection::msgs::ConnectionMsg;
     use crate::ics23_commitment::commitment::CommitmentPrefix;
     use crate::ics24_host::identifier::ClientId;
     use std::convert::TryFrom;
